+++ EcoManage.Desktop/Suppliers/FormListSuppliers.cs
@@ -11,6 +11,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 using System.Windows.Forms;
+using EcoManage.Domain.Responses;
 
 namespace EcoManage.Desktop.Suppliers
 {
